+++ src/models/build.rs
@@ -19,4 +19,11 @@ impl GenomeBuild {
             GenomeBuild::Grch38 => "38",
         }
     }
+
+    pub fn cytoband_path(&self) -> &'static str {
+        match self {
+            GenomeBuild::Grch37 => "resources/cytoBand_hg19.txt.gz",
+            GenomeBuild::Grch38 => "resources/cytoBand_hg38.txt.gz",
+        }
+    }
 }