+++ src/models/cytoband.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Clone)]
+pub struct Cytoband {
+    pub start: u64,
+    pub end: u64,
+    pub name: String,
+}